# apps/odonto-api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/treatment-plans": {
      "post": {
        "operationId": "TreatmentPlansController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTreatmentPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The treatment plan has been successfully created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new treatment plan",
        "tags": [
          "Treatment Plans"
        ]
      },
      "get": {
        "operationId": "TreatmentPlansController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all treatment plans for the clinic",
        "tags": [
          "Treatment Plans"
        ]
      }
    },
    "/treatment-plans/patient/{patientId}": {
      "get": {
        "operationId": "TreatmentPlansController_findByPatient",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all treatment plans for a patient",
        "tags": [
          "Treatment Plans"
        ]
      }
    },
    "/treatment-plans/{id}": {
      "get": {
        "operationId": "TreatmentPlansController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a treatment plan by id",
        "tags": [
          "Treatment Plans"
        ]
      },
      "patch": {
        "operationId": "TreatmentPlansController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTreatmentPlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a treatment plan",
        "tags": [
          "Treatment Plans"
        ]
      },
      "delete": {
        "operationId": "TreatmentPlansController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a treatment plan",
        "tags": [
          "Treatment Plans"
        ]
      }
    },
...
  },
  "info": {
    "title": "OdontoTec API",
    "description": "Clínica Odontológica Multi-tenant API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TreatmentPlanItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "toothNumber": {
            "type": "number"
          },
          "surface": {
            "type": "object"
          },
          "status": {
            "type": "string",
            "enum": [
              "PLANNED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELED",
              "CANCELLED"
            ]
          }
        },
        "required": [
          "description",
          "value"
        ]
      },
      "CreateTreatmentPlanDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "number"
          },
          "dentistId": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_APPROVAL",
              "APPROVED",
              "REJECTED",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "title": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TreatmentPlanItemDto"
            }
          }
        },
        "required": [
          "patientId",
          "dentistId",
          "items"
        ]
      },
      "UpdateTreatmentPlanDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "number"
          },
          "dentistId": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_APPROVAL",
              "APPROVED",
              "REJECTED",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "title": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TreatmentPlanItemDto"
            }
          }
        }
      },
...
    }
  }
}